+++ packages/toolkit/plugin-v2/src/cli/hooks.ts
@@ -1,9 +1,4 @@
 import type {
-  ModifyBundlerChainFn,
-  ModifyRsbuildConfigFn,
-  ModifyRspackConfigFn,
-  ModifyWebpackChainFn,
-  ModifyWebpackConfigFn,
   OnAfterBuildFn,
   OnAfterCreateCompilerFn,
   OnBeforeBuildFn,
@@ -14,9 +9,14 @@ import type {
   AddCommandFn,
   AddWatchFilesFn,
   ConfigFn,
+  ModifyBundlerChainFn,
   ModifyConfigFn,
   ModifyHtmlPartialsFn,
   ModifyResolvedConfigFn,
+  ModifyRsbuildConfigFn,
+  ModifyRspackConfigFn,
+  ModifyWebpackChainFn,
+  ModifyWebpackConfigFn,
   OnAfterDeployFn,
   OnAfterDevFn,
   OnBeforeDeployFn,
@@ -28,7 +28,7 @@ import type {
 } from '../types/cli/hooks';
 import type { DeepPartial } from '../types/utils';
 
-export function initHooks<Config, NormalizedConfig>() {
+export function initHooks<Config, NormalizedConfig, ExtendBuildUtils>() {
   return {
     /**
      * add config for this cli plugin
@@ -45,11 +45,16 @@ export function initHooks<Config, NormalizedConfig>() {
     modifyResolvedConfig:
       createAsyncHook<ModifyResolvedConfigFn<NormalizedConfig>>(),
 
-    modifyRsbuildConfig: createAsyncHook<ModifyRsbuildConfigFn>(),
-    modifyBundlerChain: createAsyncHook<ModifyBundlerChainFn>(),
-    modifyRspackConfig: createAsyncHook<ModifyRspackConfigFn>(),
-    modifyWebpackChain: createAsyncHook<ModifyWebpackChainFn>(),
-    modifyWebpackConfig: createAsyncHook<ModifyWebpackConfigFn>(),
+    modifyRsbuildConfig:
+      createAsyncHook<ModifyRsbuildConfigFn<ExtendBuildUtils>>(),
+    modifyBundlerChain:
+      createAsyncHook<ModifyBundlerChainFn<ExtendBuildUtils>>(),
+    modifyRspackConfig:
+      createAsyncHook<ModifyRspackConfigFn<ExtendBuildUtils>>(),
+    modifyWebpackChain:
+      createAsyncHook<ModifyWebpackChainFn<ExtendBuildUtils>>(),
+    modifyWebpackConfig:
+      createAsyncHook<ModifyWebpackConfigFn<ExtendBuildUtils>>(),
     modifyHtmlPartials: createAsyncHook<ModifyHtmlPartialsFn>(),
 
     addCommand: createAsyncHook<AddCommandFn>(),
@@ -70,6 +75,6 @@ export function initHooks<Config, NormalizedConfig>() {
   };
 }
 
-export type Hooks<Config, NormalizedConfig> = ReturnType<
-  typeof initHooks<Config, NormalizedConfig>
+export type Hooks<Config, NormalizedConfig, ExtendBuildUtils> = ReturnType<
+  typeof initHooks<Config, NormalizedConfig, ExtendBuildUtils>
 >;