+++ packages/toolkit/plugin-v2/src/types/cli/api.ts
@@ -1,25 +1,25 @@
 import type {
-  ModifyBundlerChainFn,
-  ModifyRsbuildConfigFn,
-  ModifyRspackConfigFn,
-  ModifyWebpackChainFn,
-  ModifyWebpackConfigFn,
   OnAfterBuildFn,
   OnAfterCreateCompilerFn,
   OnBeforeBuildFn,
   OnBeforeCreateCompilerFn,
 } from '@rsbuild/core';
 import type { Hooks } from '../../cli/hooks';
-import type { PluginHook, PluginHookTap } from '../hooks';
+import type { PluginHookTap } from '../hooks';
 import type { DeepPartial } from '../utils';
 import type { AppContext } from './context';
 import type {
   AddCommandFn,
   AddWatchFilesFn,
   ConfigFn,
+  ModifyBundlerChainFn,
   ModifyConfigFn,
   ModifyHtmlPartialsFn,
   ModifyResolvedConfigFn,
+  ModifyRsbuildConfigFn,
+  ModifyRspackConfigFn,
+  ModifyWebpackChainFn,
+  ModifyWebpackConfigFn,
   OnAfterDeployFn,
   OnAfterDevFn,
   OnBeforeDeployFn,
@@ -35,11 +35,16 @@ import type { CLIPluginExtends } from './plugin';
  * Define a generic CLI plugin API that provider can extend as needed.
  */
 export type CLIPluginAPI<Extends extends CLIPluginExtends> = Readonly<{
+  isPluginExists: (name: string) => boolean;
   getAppContext: () => Readonly<AppContext<Extends> & Extends['extendContext']>;
   getConfig: () => Readonly<Extends['config']>;
   getNormalizedConfig: () => Readonly<Extends['normalizedConfig']>;
   getHooks: () => Readonly<
-    Hooks<Extends['config'], Extends['normalizedConfig']> &
+    Hooks<
+      Extends['config'],
+      Extends['normalizedConfig'],
+      Extends['extendBuildUtils']
+    > &
       Extends['extendHooks']
   >;
 
@@ -55,14 +60,24 @@ export type CLIPluginAPI<Extends extends CLIPluginExtends> = Readonly<{
   >;
 
   // modify rsbuild config hooks
-  modifyRsbuildConfig: PluginHookTap<ModifyRsbuildConfigFn>;
-  modifyBundlerChain: PluginHookTap<ModifyBundlerChainFn>;
+  modifyRsbuildConfig: PluginHookTap<
+    ModifyRsbuildConfigFn<Extends['extendBuildUtils']>
+  >;
+  modifyBundlerChain: PluginHookTap<
+    ModifyBundlerChainFn<Extends['extendBuildUtils']>
+  >;
   /** Only works when bundler is Rspack */
-  modifyRspackConfig: PluginHookTap<ModifyRspackConfigFn>;
+  modifyRspackConfig: PluginHookTap<
+    ModifyRspackConfigFn<Extends['extendBuildUtils']>
+  >;
   /** Only works when bundler is Webpack */
-  modifyWebpackChain: PluginHookTap<ModifyWebpackChainFn>;
+  modifyWebpackChain: PluginHookTap<
+    ModifyWebpackChainFn<Extends['extendBuildUtils']>
+  >;
   /** Only works when bundler is Webpack */
-  modifyWebpackConfig: PluginHookTap<ModifyWebpackConfigFn>;
+  modifyWebpackConfig: PluginHookTap<
+    ModifyWebpackConfigFn<Extends['extendBuildUtils']>
+  >;
   modifyHtmlPartials: PluginHookTap<ModifyHtmlPartialsFn>;
 
   addCommand: PluginHookTap<AddCommandFn>;