+++ tests/integration/ssg/fixtures/mega-list-routes/src/routes/page.jsx
@@ -1,0 +1,9 @@
+const App = () => {
+  return (
+    <div className="text-center" id="data">
+      Hello, Home
+    </div>
+  );
+};
+
+export default App;