+++ tests/integration/ssg/fixtures/mega-list-routes/src/routes/user.[id]/page.data.js
@@ -1,0 +1,6 @@
+// Test that this file is not imported into the client bundle
+const fs = require('fs');
+
+export const loader = () => {
+  return 'Hello, User';
+};