+++ packages/cli/plugin-bff/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @modern-js/plugin-bff
 
+## 2.63.6
+
+### Patch Changes
+
+- Updated dependencies [a7a4573]
+  - @modern-js/server-core@2.63.6
+  - @modern-js/server-utils@2.63.6
+  - @modern-js/bff-core@2.63.6
+  - @modern-js/create-request@2.63.6
+  - @modern-js/utils@2.63.6
+
+## 2.63.5
+
+### Patch Changes
+
+- @modern-js/bff-core@2.63.5
+- @modern-js/server-core@2.63.5
+- @modern-js/create-request@2.63.5
+- @modern-js/server-utils@2.63.5
+- @modern-js/utils@2.63.5
+
 ## 2.63.4
 
 ### Patch Changes