+++ packages/cli/plugin-data-loader/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @modern-js/plugin-data-loader
 
+## 2.63.6
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.63.6
+- @modern-js/utils@2.63.6
+
+## 2.63.5
+
+### Patch Changes
+
+- @modern-js/runtime-utils@2.63.5
+- @modern-js/utils@2.63.5
+
 ## 2.63.4
 
 ### Patch Changes