+++ packages/cli/plugin-tailwind/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @modern-js/plugin-tailwindcss
 
+## 2.63.6
+
+### Patch Changes
+
+- Updated dependencies [c3e73b4]
+  - @modern-js/runtime@2.63.6
+  - @modern-js/node-bundle-require@2.63.6
+  - @modern-js/runtime-utils@2.63.6
+  - @modern-js/utils@2.63.6
+
+## 2.63.5
+
+### Patch Changes
+
+- @modern-js/runtime@2.63.5
+- @modern-js/node-bundle-require@2.63.5
+- @modern-js/runtime-utils@2.63.5
+- @modern-js/utils@2.63.5
+
 ## 2.63.4
 
 ### Patch Changes