+++ packages/document/main-doc/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/main-doc
 
+## 2.63.6
+
+### Patch Changes
+
+- 5fd59c6: feat: deprecate beforeConfig hook
+
+  feat: 废弃 beforeConfig Hook
+
+  - @modern-js/sandpack-react@2.63.6
+
+## 2.63.5
+
+### Patch Changes
+
+- @modern-js/sandpack-react@2.63.5
+
 ## 2.63.4
 
 ### Patch Changes