+++ packages/generator/generator-cases/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/generator-cases
 
+## 3.7.12
+
+### Patch Changes
+
+- @modern-js/generator-common@3.7.12
+
+## 3.7.11
+
+### Patch Changes
+
+- @modern-js/generator-common@3.7.11
+
 ## 3.7.10
 
 ### Patch Changes