+++ packages/generator/generators/server-generator/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @modern-js/server-generator
 
+## 3.7.12
+
+## 3.7.11
+
 ## 3.7.10
 
 ## 3.7.9