+++ packages/runtime/plugin-runtime/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @modern-js/runtime
 
+## 2.63.6
+
+### Patch Changes
+
+- c3e73b4: fix: add missing getInitData api
+  fix: 添加缺失的 getInitData 接口
+- Updated dependencies [4492a5d]
+  - @modern-js/plugin-v2@2.63.6
+  - @modern-js/plugin-data-loader@2.63.6
+  - @modern-js/plugin@2.63.6
+  - @modern-js/runtime-utils@2.63.6
+  - @modern-js/types@2.63.6
+  - @modern-js/utils@2.63.6
+
+## 2.63.5
+
+### Patch Changes
+
+- @modern-js/plugin-v2@2.63.5
+- @modern-js/plugin-data-loader@2.63.5
+- @modern-js/plugin@2.63.5
+- @modern-js/runtime-utils@2.63.5
+- @modern-js/types@2.63.5
+- @modern-js/utils@2.63.5
+
 ## 2.63.4
 
 ### Patch Changes