+++ packages/server/babel-plugin-module-resolver/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @modern-js/babel-plugin-module-resolver
 
+## 2.63.6
+
+## 2.63.5
+
 ## 2.63.4
 
 ## 2.63.3