+++ packages/server/plugin-koa/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @modern-js/plugin-koa
 
+## 2.63.6
+
+### Patch Changes
+
+- Updated dependencies [a7a4573]
+  - @modern-js/server-core@2.63.6
+  - @modern-js/bff-core@2.63.6
+  - @modern-js/bff-runtime@2.63.6
+  - @modern-js/types@2.63.6
+  - @modern-js/utils@2.63.6
+
+## 2.63.5
+
+### Patch Changes
+
+- @modern-js/bff-core@2.63.5
+- @modern-js/bff-runtime@2.63.5
+- @modern-js/server-core@2.63.5
+- @modern-js/types@2.63.5
+- @modern-js/utils@2.63.5
+
 ## 2.63.4
 
 ### Patch Changes