+++ packages/server/plugin-worker/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/plugin-worker
 
+## 2.63.6
+
+### Patch Changes
+
+- @modern-js/prod-server@2.63.6
+- @modern-js/server-utils@2.63.6
+- @modern-js/utils@2.63.6
+
+## 2.63.5
+
+### Patch Changes
+
+- @modern-js/prod-server@2.63.5
+- @modern-js/server-utils@2.63.5
+- @modern-js/utils@2.63.5
+
 ## 2.63.4
 
 ### Patch Changes