+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @modern-js/server-utils
 
+## 2.63.6
+
+### Patch Changes
+
+- @modern-js/babel-preset@2.63.6
+- @modern-js/babel-plugin-module-resolver@2.63.6
+- @modern-js/babel-compiler@2.63.6
+- @modern-js/utils@2.63.6
+
+## 2.63.5
+
+### Patch Changes
+
+- @modern-js/babel-preset@2.63.5
+- @modern-js/babel-plugin-module-resolver@2.63.5
+- @modern-js/babel-compiler@2.63.5
+- @modern-js/utils@2.63.5
+
 ## 2.63.4
 
 ### Patch Changes